Fold ~PortalTheme to Root
