+++ packages/mui/src/Root.tsx
@@ -4,9 +4,14 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as React from "react";
-import { ThemeProvider, useColorScheme } from "@mui/material/styles";
+import {
+	type Theme,
+	ThemeProvider,
+	useColorScheme,
+} from "@mui/material/styles";
 import { Root as StrataKitRoot } from "@stratakit/foundations";
 import {
+	PortalContext,
 	PortalWrapperContext,
 	RootContext,
 } from "@stratakit/foundations/secret-internals";
@@ -14,7 +19,6 @@ import { useSafeContext } from "@stratakit/internal-utils/hooks";
 import { forwardRef } from "@stratakit/internal-utils/react";
 import cx from "classnames";
 import { createTheme } from "./~createTheme.js";
-import { PortalTheme } from "./~PortalTheme.js";
 import { StyledEngineProvider } from "./Root.internal.js";
 import css from "./styles.css.js";
 
@@ -82,12 +86,6 @@ DEV: Root.displayName = "Root";
 
 // ----------------------------------------------------------------------------
 
-function wrapPortal(portal: React.ReactNode) {
-	return <PortalTheme>{portal}</PortalTheme>;
-}
-
-// ----------------------------------------------------------------------------
-
 interface RootInnerProps
 	extends BaseProps<"div">,
 		Pick<RootProps, "colorScheme" | "unstable_accentColor" | "rootNode"> {}
@@ -145,6 +143,55 @@ function Styles() {
 
 	return null;
 }
+
+// ----------------------------------------------------------------------------
+
+function wrapPortal(portal: React.ReactNode) {
+	return <PortalThemeProvider>{portal}</PortalThemeProvider>;
+}
+
+// ----------------------------------------------------------------------------
+
+function PortalThemeProvider(props: React.PropsWithChildren) {
+	const { container: containerEl, unstable_getContainer } =
+		useSafeContext(PortalContext);
+	const container = unstable_getContainer ?? containerEl;
+	const theme = React.useCallback(
+		(outerTheme: Theme): Theme => {
+			return {
+				...outerTheme,
+				components: {
+					...outerTheme.components,
+					MuiModal: {
+						...outerTheme.components?.MuiModal,
+						defaultProps: {
+							...outerTheme.components?.MuiModal?.defaultProps,
+							container,
+						},
+					},
+					MuiPopover: {
+						...outerTheme.components?.MuiPopover,
+						defaultProps: {
+							...outerTheme.components?.MuiPopover?.defaultProps,
+							// Popover passes down `container` prop to `Modal` https://github.com/mui/material-ui/blob/708ef10e874efa63d2e4972bd902befa1912f2dc/packages/mui-material/src/Popover/Popover.js#L389
+							container,
+						},
+					},
+					MuiPopper: {
+						...outerTheme.components?.MuiPopper,
+						defaultProps: {
+							...outerTheme.components?.MuiPopper?.defaultProps,
+							container,
+						},
+					},
+				},
+			};
+		},
+		[container],
+	);
+	return <ThemeProvider theme={theme} {...props} />;
+}
+DEV: PortalThemeProvider.displayName = "PortalThemeProvider";
 
 // ----------------------------------------------------------------------------
